+++ common/src/main/java/net/mehvahdjukaar/polytone/dimension/DimensionEffectsModifier.java
@@ -6,6 +6,7 @@ import com.mojang.serialization.Decoder;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.mehvahdjukaar.polytone.PlatStuff;
 import net.mehvahdjukaar.polytone.Polytone;
+import net.mehvahdjukaar.polytone.block.BlockContextExpression;
 import net.mehvahdjukaar.polytone.colormap.Colormap;
 import net.mehvahdjukaar.polytone.colormap.ColormapExpressionProvider;
 import net.mehvahdjukaar.polytone.colormap.IColorGetter;
@@ -19,7 +20,7 @@ import java.util.Optional;
 import java.util.Set;
 
 
-public record DimensionEffectsModifier(Optional<Either<Float, ColormapExpressionProvider>> cloudLevel,
+public record DimensionEffectsModifier(Optional<Either<Float, BlockContextExpression>> cloudLevel,
                                        Optional<Boolean> hasGround,
                                        Optional<DimensionSpecialEffects.SkyType> skyType,
                                        Optional<Boolean> forceBrightLightmap,
@@ -36,7 +37,7 @@ public record DimensionEffectsModifier(Optional<Either<Float, ColormapExpression
 
     public static final Decoder<DimensionEffectsModifier> CODEC = RecordCodecBuilder.create(instance ->
             instance.group(
-                    Codec.either(Codec.FLOAT, ColormapExpressionProvider.CODEC).optionalFieldOf("cloud_level").forGetter(DimensionEffectsModifier::cloudLevel),
+                    Codec.either(Codec.FLOAT, BlockContextExpression.CODEC).optionalFieldOf("cloud_level").forGetter(DimensionEffectsModifier::cloudLevel),
                     Codec.BOOL.optionalFieldOf("has_ground").forGetter(DimensionEffectsModifier::hasGround),
                     SKY_TYPE_CODEC.optionalFieldOf("sky_type").forGetter(DimensionEffectsModifier::skyType),
                     Codec.BOOL.optionalFieldOf("force_bright_lightmap").forGetter(DimensionEffectsModifier::forceBrightLightmap),
@@ -88,7 +89,7 @@ public record DimensionEffectsModifier(Optional<Either<Float, ColormapExpression
 
     public DimensionEffectsModifier applyInplace(ResourceLocation dimensionId) {
         DimensionSpecialEffects effects = PlatStuff.getDimensionEffects(dimensionId);
-        Optional<Either<Float, ColormapExpressionProvider>> oldCloud = Optional.empty();
+        Optional<Either<Float, BlockContextExpression>> oldCloud = Optional.empty();
         if (this.cloudLevel.isPresent() && this.cloudLevel.get().left().isPresent()) {
             oldCloud = Optional.of(Either.left(effects.cloudLevel));
             effects.cloudLevel = this.cloudLevel.get().left().get();